Split Android cloud remote service
